ultimate refactor (still bugs remain)

+++ src/augs/misc/delta.h
@@ -1,16 +1,17 @@
 #pragma once
+#include "augs/math/declare_math.h"
 
 namespace augs {
 	struct introspection_access;
 
 	class delta {
 		// GEN INTROSPECTOR class augs::delta
-		float delta_ms;
+		real32 delta_ms;
 		// END GEN INTROSPECTOR
 
 		friend struct introspection_access;
 	public:
-		delta(const float delta_ms = 0.f) 
+		delta(const real32 delta_ms = static_cast<real32>(0))
 			: delta_ms(delta_ms)
 		{}
